Continuing to minimize imports
