+++ functions/utils/m2m.js
@@ -1,14 +1,6 @@
 const fetch = require("node-fetch");
-const functions = require("firebase-functions");
 
-const client_vars = functions.config().client_vars;
-const auth0Config = functions.config().auth;
-
-const client_id = client_vars ? client_vars.client_id : "";
-const client_secret = client_vars ? client_vars.client_secret : "";
-const issuer = auth0Config ? auth0Config.issuer : "";
-
-const getM2MToken = () => {
+const getM2MToken = (client_id, client_secret, issuer) => {
   options = {
     method: "POST",
     headers: {
@@ -23,11 +15,9 @@ const getM2MToken = () => {
   };
   return fetch(`${issuer}oauth/token`, options)
     .then((res) => {
-      //console.log(res)
       return res.json();
     })
     .then((data) => {
-      //console.log(data)
       return data.access_token;
     })
     .catch((err) => {
